requirements.txt
streamlit
pandas
numpy
plotly
textblob
requests
tweepy
youtube-comment-downloader
python-dotenv